\name{dummy_data}
\alias{dummy_data}
\docType{data}
\title{
Abundance of marine macrobenthic invertebrates of South Australia.
}
\description{
The dataset shows dummy values of abundance of marine macrobenthic fauana of 20 different sites.
}
\usage{data("dummy_data")}
\format{
  The format is:
 chr "dummy_data"
}
\details{
Abundance presented here are only an example of how the database should be constructed.
}
\source{
Lam-Gordillo, O., Baring, R., Dittmann, S. 2020. Establishing the South Australian Macrobenthic Traits Database (SAMT): a trait classification for functional assessments.
}
\references{
Lam-Gordillo, O., Baring, R., Dittmann, S. 2020. Establishing the South Australian Macrobenthic Traits Database (SAMT): a trait classification for functional assessments.
}
\examples{
data(dummy_data)
str(dummy_data) ; head(dummy_data) ; summary(dummy_data)
}
\keyword{datasets}
