\name{SAMTdatabase}
\alias{SAMTdatabase}
\docType{data}
\title{
The South Australian Macrobenthic Traits (SAMT) Database.
}
\description{
This is a database showing the functional trait classification (fuzzy coding) of 277 marine macrobenthic taxa of soft sediments of South Australia.
}
\usage{data("SAMTdatabase")}
\format{
  The format is:
 chr "SAMTdatabase"
}
\details{
The functional trait classification ranges from 0-1, with 0 being no affinity and 1 being hight affinity to a trait.
}
\source{
Traits information was retrieved from multiple sources, see information detailed in Lam-Gordillo, O., Baring, R., Dittmann, S. 2020. Establishing the South Australian Macrobenthic Traits Database (SAMT): a trait classification for functional assessments.
}
\references{
Lam-Gordillo, O., Baring, R., Dittmann, S. 2020. Establishing the South Australian Macrobenthic Traits Database (SAMT): a trait classification for functional assessments.
}
\examples{
data(SAMTdatabase)
str(dummy_data) ; head(dummy_data)
}
\keyword{datasets}
