% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAMT.R
\name{SAMTdb}
\alias{SAMTdb}
\title{SAMTdb}
\usage{
SAMTdb()
}
\arguments{
\item{SAMTdb}{A database with information on macrobenthic taxa and their functional trait classification.}
}
\value{
A data frame with the taxa list with fuzzy coding or functional trait classification.
}
\description{
SAMTdb is a databse designed for marine macroinvertebrates taxa in South Australia (values range from 0 to 1).
}
\details{
This function provides the fuzzy coding (functional trait calssification) allocated to each marine macroinvertebrate taxa of South Australia.
}
\examples{
SAMTdb()
}
\keyword{fuzzy}
