% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAMT.R
\name{Taxa.list}
\alias{Taxa.list}
\title{Taxa list}
\usage{
Taxa.list()
}
\arguments{
\item{Taxa.list}{Returns a dataframe with a list of taxa names.}
}
\value{
A data frame with the Taxa list.
}
\description{
Taxa_list provides a list of names of marine macroinvertebrates taxa contained in SAMTdatabase.
}
\details{
This function provides the marine macroinvertebrates list from soft sediments in South Australia, available in SAMT database.
}
\examples{
Taxa.list()
}
\keyword{taxa}
