% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAMT.R
\name{Traits.list}
\alias{Traits.list}
\title{Traits list}
\usage{
Traits.list()
}
\arguments{
\item{Traits.list}{Returns a data frame with names of the traits used in the SAMTdatabase.}

\item{Traits}{A full list of traits and traits modalities.}
}
\value{
A data frame with the traits list.
}
\description{
List of traits and trait-modalities used for classifying marine macroinvertebrates taxa of soft sediments in South Australia.
}
\details{
This function provides the list of traits available in SAMT database.
}
\examples{
Traits.list()
}
\keyword{traits}
