\name{Traits}
\alias{Traits}
\docType{data}
\title{
Traits.
}
\description{
This dataset shows a list of the traits used in SAMT databese.
}
\usage{data("Traits")}
\format{
  The format is:
 chr "Traits"
}
\details{
The traits list shows 13 traits and 54 traits-modalities.
}
\source{
Lam-Gordillo, O., Baring, R., Dittmann, S. 2020. Ecosystem functioning and functional approaches on marine macrobenthic fauna: A research synthesis towards a global concensus. Ecological Indicators. https://doi.org/10.1016/j.ecolind.2020.106379
}
\references{
Lam-Gordillo, O., Baring, R., Dittmann, S. 2020. Establishing the South Australian Macrobenthic Traits Database (SAMT): a trait classification for functional assessments.
}
\examples{
data(Traits)
}
\keyword{datasets}
