% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAMT.R
\name{TVSR}
\alias{TVSR}
\title{Traits Values per Site/Replicate}
\usage{
TVSR(a, b)
}
\arguments{
\item{a}{A database or datamatrix that contains taxa abundace per site/replicate. In all cases, taxa labels are requiered to match with the traits database.}

\item{b}{SAMTdatabse (recommended), but it could be another database or datamatrix that contains the same number of functional traits by taxa that SAMTdatabase. In all cases, taxa labels are requiered to match with the abundance database.}

\item{TVSR}{Name of the function.}
}
\value{
A data frame with the Trait Values for each Site/Replicate//location/time (TVSR).
}
\description{
TVSR calculate the trait value by site/replicate level (Functional classification * Taxa Abudance(specific location/time)).
}
\details{
This function calcualte the traits values for specific site / replicate / region / time.
}
\examples{
TVSR(dummy_data,SAMTdatabase)
}
\keyword{TVSR}
