% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SAMT.R
\name{TVT}
\alias{TVT}
\title{Traits values by taxa}
\usage{
TVT(a, b)
}
\arguments{
\item{a}{A database or datamatrix that contains taxa abundace per site/replicate. In all cases, taxa labels are requiered to match with the traits database.}

\item{b}{SAMTdatabse (recommended), but it could be another database or datamatrix that contains the same number of functional traits by taxa that SAMTdatabase. In all cases, taxa labels are requiered to match with the abundance database.}

\item{TVT}{Name of the function.}
}
\value{
A data frame with the Trait Values by each Taxa (TVT).
}
\description{
TVT calculate trait values at taxa level taking in consideration the taxa abundance (Funtional clasiffication by Taxa abudance).
}
\details{
This function multiply the trait values per each taxa by the abundance per each taxa.
}
\examples{
TVT(dummy_data,SAMTdatabase)
}
\keyword{TVT}
